requirements.txt
Pillow
NumPy
SciPy
Matplotlib
OpenCV-Python
Scikit-learn